remove the drop guard of Dirty
If a struct is constructed with `Dirty::new_dirty`, any following
returned errors will cause an assertion failure when dropping it,
because the struct is dirty, it is too complicated to clear the dirty
state in every subsequent error handler.

After this commit, users **MUST** ensure that the useful dirty struct is
correctly flushed before dropping it. If the dirty struct becomes useless
because of errors, users can just drop it.